lets try to work with courant servers

+++ student/sft_experiments.py
@@ -50,7 +50,7 @@ def init_vllm(model_id, device, seed, gpu_memory_utilization=0.85):
     from vllm.model_executor import set_random_seed as vllm_set_random_seed
     vllm_set_random_seed(seed)
     world_size_patch = patch("torch.distributed.get_world_size", return_value=1)
-    profiling_patch  = patch(
+    profiling_patch = patch(
         "vllm.worker.worker.Worker._assert_memory_footprint_increased_during_profiling",
         return_value=None,
     )
@@ -58,7 +58,7 @@ def init_vllm(model_id, device, seed, gpu_memory_utilization=0.85):
         return LLM(
             model=model_id,
             device=device,
-            dtype=torch.bfloat16,
+            dtype=torch.float16, 
             enable_prefix_caching=True,
             gpu_memory_utilization=gpu_memory_utilization,
         )
@@ -93,7 +93,7 @@ def train(args):
     tokenizer = AutoTokenizer.from_pretrained(args.model, trust_remote_code=True)
     policy = AutoModelForCausalLM.from_pretrained(
         args.model,
-        torch_dtype=torch.bfloat16,
+        torch_dtype=torch.float16,
         trust_remote_code=True,
     ).to(device)
